refactor(rename): arbitrarily rename utils.js to helpers.js
Because I prefer it now 😄
