+++ server/config.js
@@ -1,4 +1,4 @@
-import { assertNoUndefinedValues } from './utils.js';
+import { assertNoUndefinedValues } from './helpers.js';
 
 const config = {
 	API_URL: process.env.API_URL || 'https://pokeapi.co/api/v2',
