+++ server/init.js
@@ -1,6 +1,6 @@
 import config from './config.js';
 import app from './app.js';
-import { logger } from './utils.js';
+import { logger } from './helpers.js';
 
 const { PORT } = config;
